Moved GPS Switch, improved create group button styling, cleaned up styles
