+++ src/app/components/GPSSwitch/styles.js
@@ -1,0 +1,43 @@
+
+import {PURPLE, LIGHT_GREY_2, DARK_PURPLE} from '../../common/colors';
+
+export default {
+  rocketIcon: {
+    fontSize: 30,
+    color: DARK_PURPLE
+  },
+  innerCircleStyle:{
+    alignItems: 'center',
+    justifyContent: 'center',
+    borderWidth: 0,
+    borderColor: 'white'
+  },
+  outerCircleStyle:{
+  },
+  activeText: {
+    color: DARK_PURPLE,
+    fontSize: 18,
+    fontWeight: 'bold',
+  },
+  plus: {
+    color: PURPLE,
+    paddingTop: 10
+  },
+  memberWrapper: {
+    paddingBottom: 15,
+    borderColor: LIGHT_GREY_2,
+    borderBottomWidth: 1
+  },
+  inActiveText: {
+    color: DARK_PURPLE,
+    fontSize: 18,
+    fontWeight: 'bold',
+  },
+  text: {
+    color: DARK_PURPLE,
+    fontSize: 18,
+  },
+  containerRelative:{
+    position: 'relative',
+  }
+};
