+++ src/app/containers/Map/actions.js
@@ -27,12 +27,3 @@ export function addParticipantsToMap({mapID, participants = []}) {
     participants
   };
 }
-
-export function mapGPSToggled({mapID, gpsEnabled}) {
-  return {
-    type: ActionTypes.MAP_GPS_TOGGLED,
-    mapID,
-    gpsEnabled,
-    gpsEnabledAt: (new Date().getTime())
-  };
-}
